[Android/Feat] blockChain......

See merge request s12-blockchain-nft-sub1/S12P21D210!238

+++ Android/app/src/main/java/com/example/second_project/data/model/dto/request/DepositRequest.kt
@@ -1,6 +1,8 @@
 package com.example.second_project.data.model.dto.request
 
+import java.math.BigInteger
+
 data class DepositRequest(
     val userId: Int,
-    val quantity: Int
+    val quantity: BigInteger
 )
